+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-02T06:41:47.199Z",
+    "2026-09-02T07:25:06.598Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {