+++ packages/orq-rc/RELEASES.md
@@ -12019,3 +12019,13 @@ Based on:
 - [typescript v4.15.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.12 - packages/orq-rc
+
+## 2026-09-02 07:28:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.2 (2.933.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.13 - packages/orq-rc